streamlit2/requirements.txt
onnxruntime
numpy
pillow
tqdm
pickle-mixin
streamlit
psycopg2-binary
pgvector
scikit-learn
imagehash
transformers 
torch 
torchvision 
torchaudio